requirements.txt
pyperclip
Pillow
PyYAML
oss2
qiniu
cos-python-sdk-v5
